+++ src/features/videos/index.js
@@ -1,0 +1,5 @@
+/**
+ * Videos Feature Module
+ */
+
+export { default as VideoVault } from './VideoVault'